Unit test suite refactors and new storyboards proposal
