--- app/services/tests/test_activity_logger.py
@@ -1,164 +1,0 @@
-import pytest
-from unittest.mock import ANY
-from gundi_core.events import (
-    LogLevel,
-    IntegrationActionStarted,
-    IntegrationActionComplete,
-    IntegrationActionFailed,
-    IntegrationActionCustomLog
-)
-from app import settings
-from app.services.activity_logger import publish_event, activity_logger, log_activity
-
-
-@pytest.mark.parametrize(
-    "system_event",
-    ["action_started_event", "action_complete_event", "action_failed_event", "custom_activity_log_event"],
-    indirect=["system_event"])
-@pytest.mark.asyncio
-async def test_publish_event(
-        mocker, mock_pubsub_client, integration_event_pubsub_message, gcp_pubsub_publish_response,
-        system_event
-):
-    mocker.patch("app.services.activity_logger.pubsub", mock_pubsub_client)
-
-    response = await publish_event(
-        event=system_event,
-        topic_name=settings.INTEGRATION_EVENTS_TOPIC
-    )
-
-    assert response == gcp_pubsub_publish_response
-    assert mock_pubsub_client.PublisherClient.called
-    assert mock_pubsub_client.PubsubMessage.called
-    assert mock_pubsub_client.PublisherClient.called
-    assert mock_pubsub_client.PublisherClient.return_value.publish.called
-    mock_pubsub_client.PublisherClient.return_value.publish.assert_any_call(
-        f"projects/{settings.GCP_PROJECT_ID}/topics/{settings.INTEGRATION_EVENTS_TOPIC}",
-        [integration_event_pubsub_message],
-    )
-
-
-@pytest.mark.asyncio
-async def test_activity_logger_decorator(
-        mocker, mock_publish_event, integration_v2, pull_observations_config
-):
-
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-
-    @activity_logger()
-    async def action_pull_observations(integration, action_config):
-        return {"observations_extracted": 10}
-
-    await action_pull_observations(
-        integration=integration_v2,
-        action_config=pull_observations_config
-    )
-
-    # Two events expected: One on start and one on completion
-    assert mock_publish_event.call_count == 2
-    assert isinstance(mock_publish_event.call_args_list[0].kwargs.get("event"), IntegrationActionStarted)
-    assert isinstance(mock_publish_event.call_args_list[1].kwargs.get("event"), IntegrationActionComplete)
-
-
-@pytest.mark.asyncio
-async def test_activity_logger_decorator_with_arguments(
-        mocker, mock_publish_event, integration_v2, pull_observations_config
-):
-
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-
-    @activity_logger(on_start=False, on_completion=True, on_error=False)
-    async def action_pull_observations(integration, action_config):
-        return {"observations_extracted": 10}
-
-    await action_pull_observations(
-        integration=integration_v2,
-        action_config=pull_observations_config
-    )
-
-    # Only one event expected, on completion
-    assert mock_publish_event.call_count == 1
-    assert isinstance(mock_publish_event.call_args_list[0].kwargs.get("event"), IntegrationActionComplete)
-
-
-@pytest.mark.asyncio
-async def test_activity_logger_decorator_on_error(
-        mocker, mock_publish_event, integration_v2, pull_observations_config
-):
-
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-
-    @activity_logger()
-    async def action_pull_observations(integration, action_config):
-        raise Exception("Something went wrong")
-
-    with pytest.raises(Exception):
-        await action_pull_observations(
-            integration=integration_v2,
-            action_config=pull_observations_config
-        )
-
-    # Two events expected: One on start and one on error
-    assert mock_publish_event.call_count == 2
-    assert isinstance(mock_publish_event.call_args_list[0].kwargs.get("event"), IntegrationActionStarted)
-    assert isinstance(mock_publish_event.call_args_list[1].kwargs.get("event"), IntegrationActionFailed)
-
-
-@pytest.mark.asyncio
-async def test_log_activity_with_debug_level(mocker, integration_v2, pull_observations_config, mock_publish_event):
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    await log_activity(
-        integration_id=integration_v2.id,
-        action_id="pull_observations",
-        level=LogLevel.DEBUG,
-        title="Extracted 10 observations from 2 devices",
-        data={"devices": ["deviceid1", "deviceid2"]},
-        config_data=pull_observations_config.data
-    )
-    assert mock_publish_event.call_count == 1
-    assert isinstance(mock_publish_event.call_args_list[0].kwargs.get("event"), IntegrationActionCustomLog)
-
-
-@pytest.mark.asyncio
-async def test_log_activity_with_info_level(mocker, integration_v2, mock_publish_event, pull_observations_config):
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    await log_activity(
-        integration_id=integration_v2.id,
-        action_id="pull_observations",
-        level=LogLevel.INFO,
-        title="Extracting observations with filter..",
-        data={"start_date": "2024-01-01", "end_date": "2024-01-31"},
-        config_data=pull_observations_config.data
-    )
-    assert mock_publish_event.call_count == 1
-    assert isinstance(mock_publish_event.call_args_list[0].kwargs.get("event"), IntegrationActionCustomLog)
-
-
-@pytest.mark.asyncio
-async def test_log_activity_with_warning_level(mocker, integration_v2, mock_publish_event, pull_observations_config):
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    await log_activity(
-        integration_id=integration_v2.id,
-        action_id="pull_observations",
-        level=LogLevel.WARNING,
-        title="Skipping end_date because it's greater than today. Please review your configuration.",
-        config_data=pull_observations_config.data
-    )
-    assert mock_publish_event.call_count == 1
-    assert isinstance(mock_publish_event.call_args_list[0].kwargs.get("event"), IntegrationActionCustomLog)
-
-
-@pytest.mark.asyncio
-async def test_log_activity_with_error_level(mocker, integration_v2, mock_publish_event, pull_observations_config):
-    mocker.patch("app.services.activity_logger.publish_event", mock_publish_event)
-    await log_activity(
-        integration_id=integration_v2.id,
-        action_id="pull_observations",
-        level=LogLevel.ERROR,
-        title="Error getting data from System X",
-        data={"error": "Connection error with host 'systemx.com'"},
-        config_data=pull_observations_config.data
-    )
-    assert mock_publish_event.call_count == 1
-    assert isinstance(mock_publish_event.call_args_list[0].kwargs.get("event"), IntegrationActionCustomLog)
-
